circ: Remove generated files and fix lint findings.
Drop circ's per-entry .clang-format/.clang-tidy/.gitignore in favor of
shared root configs (add .clang-tidy), remove committed tests/results
output, fix realloc/widening/strtol/malloc findings, prune comments, and
add a lint step to CI.

+++ circ/src/arena.h
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#define ARENA_BLOCK_SIZE (128 * 1024) /* 128KB per block */
+#define ARENA_BLOCK_SIZE ((size_t)128 * 1024) /* 128KB per block */
 
 typedef struct arena_block
 {
